port::UInt16

+++ src/notebook/Events.jl
@@ -49,7 +49,7 @@ end
 # Triggered when the web server gets started
 struct ServerStartEvent <: PlutoEvent
     address::String
-    port::Int
+    port::UInt16
 end
 
 # Triggered when a notebook is saved
